code/requirements.txt
requests
beautifulsoup4
pandas
openpyxl
google-genai
pypdf
groq
fastapi
uvicorn
python-multipart
python-dotenv
python-docx
jinja2
